Extract a `RequirementsResolver` class from `Test::Runner`
Plus some very minor changes. But basically this is just a refactoring
to further break down the unwieldy large `Test::Runner` class.
