+++ bin/test/tasks/build_fixtures.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::BuildFixtures < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_rake_task('spec:fixture_builder:rebuild')
   end
