+++ bin/test/tasks/compile_java_script.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::CompileJavaScript < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_system_command('bin/webpack --silent')
   end
