+++ bin/test/tasks/exit.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::Exit < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     puts("Exiting with code #{Test::Runner.exit_code}")
     exit(Test::Runner.exit_code)
