+++ bin/test/tasks/run_annotate.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::RunAnnotate < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_system_command(<<~COMMAND)
       bin/annotate --models --show-indexes --sort --exclude fixtures,tests && git diff --exit-code
