+++ bin/test/tasks/run_brakeman.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::RunBrakeman < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_system_command('bin/brakeman --quiet')
   end
