+++ bin/test/tasks/run_eslint.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::RunEslint < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_system_command(<<~COMMAND)
       ./node_modules/.bin/eslint --max-warnings 0 --ext .js,.vue app/javascript/
