+++ bin/test/tasks/run_stylelint.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::RunStylelint < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_system_command('./node_modules/.bin/stylelint app/**/*.{css,scss,vue} --max-warnings 0')
   end
