+++ bin/test/tasks/yarn_install.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Test::Tasks::YarnInstall < Pallets::Task
+  include Test::TaskHelpers
+
   def run
     execute_system_command('yarn install')
   end
